refactor: remove SwapModalHeaderMod
